Chore(.user): move package
