+++ src/main/java/com/newbarams/ajaja/global/security/common/CustomUserDetailService.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Component;
 import com.newbarams.ajaja.global.common.error.ErrorCode;
 import com.newbarams.ajaja.global.common.exception.AjajaException;
 import com.newbarams.ajaja.module.user.domain.User;
-import com.newbarams.ajaja.module.user.domain.UserRepository;
+import com.newbarams.ajaja.module.user.domain.repository.UserRepository;
 
 import lombok.RequiredArgsConstructor;
