+++ src/main/java/com/newbarams/ajaja/module/user/application/LoginService.java
@@ -7,7 +7,7 @@ import com.newbarams.ajaja.global.security.jwt.util.JwtGenerator;
 import com.newbarams.ajaja.module.user.auth.model.AccessToken;
 import com.newbarams.ajaja.module.user.auth.model.Profile;
 import com.newbarams.ajaja.module.user.domain.User;
-import com.newbarams.ajaja.module.user.domain.UserRepository;
+import com.newbarams.ajaja.module.user.domain.repository.UserRepository;
 import com.newbarams.ajaja.module.user.dto.UserResponse;
 
 import lombok.RequiredArgsConstructor;
