+++ src/main/java/com/newbarams/ajaja/module/user/domain/repository/UserRepository.java
@@ -1,8 +1,10 @@
-package com.newbarams.ajaja.module.user.domain;
+package com.newbarams.ajaja.module.user.domain.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.newbarams.ajaja.module.user.domain.User;
 
 public interface UserRepository extends JpaRepository<User, Long> {
 	Optional<User> findByEmail(String email);
