+++ src/main/java/com/newbarams/ajaja/module/user/presentation/UserController.java
@@ -2,7 +2,6 @@ package com.newbarams.ajaja.module.user.presentation;
 
 import static org.springframework.http.HttpStatus.*;
 
-import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
