+++ src/test/java/com/newbarams/ajaja/module/plan/application/UpdatePlanServiceTest.java
@@ -20,7 +20,7 @@ import com.newbarams.ajaja.module.plan.domain.RemindInfo;
 import com.newbarams.ajaja.module.plan.dto.PlanRequest;
 import com.newbarams.ajaja.module.plan.repository.PlanRepository;
 import com.newbarams.ajaja.module.user.domain.User;
-import com.newbarams.ajaja.module.user.domain.UserRepository;
+import com.newbarams.ajaja.module.user.domain.repository.UserRepository;
 
 @SpringBootTest
 class UpdatePlanServiceTest {
