+++ src/test/java/com/newbarams/ajaja/module/user/application/LogoutServiceTest.java
@@ -11,7 +11,7 @@ import org.mockito.Mock;
 import com.newbarams.ajaja.common.MockTestSupport;
 import com.newbarams.ajaja.global.common.exception.AjajaException;
 import com.newbarams.ajaja.global.security.jwt.util.JwtRemover;
-import com.newbarams.ajaja.module.user.domain.UserRepository;
+import com.newbarams.ajaja.module.user.domain.repository.UserRepository;
 
 class LogoutServiceTest extends MockTestSupport {
 	@InjectMocks
